refactor: colocate contributor selection
